5.3.4

+++ br/pkg/lightning/common/storage_unix.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build !windows
+// +build !windows
 
 // TODO: Deduplicate this implementation with DM!
 
@@ -56,7 +56,7 @@ func GetStorageSize(dir string) (size StorageSize, err error) {
 	}
 
 	// Available blocks * size per block = available space in bytes
-	size.Available = stat.Bavail * bSize
+	size.Available = uint64(stat.Bavail) * bSize
 	size.Capacity = stat.Blocks * bSize
 
 	return
